Change unit test framework to CppUTest
Introduce CppUMock mocking framework and remove BitLoom Costumes
Update configuration for Travis builds
Change structure of CMakefiles
Update README file and file header information
