+++ .travis.yml
@@ -4,13 +4,18 @@ install:
   - pip install --user cmake
 
 before_script:
-  - git clone https://github.com/ThrowTheSwitch/Unity.git ~/Unity
-  - git clone https://github.com/bluezephyr/bitloom-costumes.git ~/bitloom-costumes
+  - git clone https://github.com/cpputest/cpputest.git ~/cpputest
+  - cd ~/cpputest
+  - ./autogen.sh
+  - cd cpputest_build
+  - ../configure
+  - make -j
+  - cd ${TRAVIS_BUILD_DIR}
   - mkdir build
   - cd build
 
 script:
-  - cmake -D BITLOOM_ROOT=~ ..
-  - make
-  - ./bitloom-core-test -v
+  - cmake -DCOMPILE_TESTS=ON -DCPPUTEST_HOME=~/cpputest ..
+  - make all -j
+  - make test
