+++ src/i2c_master/i2c_master.c
@@ -4,7 +4,7 @@
  * This module implements a simple I2C master driver.  Note that interrupts
  * are not used, the communication is done using busy wait.
  *
- * Copyright (c) 2016-2018. BlueZephyr
+ * Copyright (c) 2016-2020. BlueZephyr
  *
  * This software may be modified and distributed under the terms
  * of the MIT license.  See the LICENSE file for details.
