+++ src/scheduler/CMakeLists.txt
@@ -1,12 +1,10 @@
-cmake_minimum_required(VERSION 3.12)
-project(scheduler C)
-
+enable_language(C)
 set(CMAKE_C_STANDARD 99)
 
 add_library(scheduler
-        scheduler.c
-        ${BITLOOM_ROOT}/bitloom-core/include/core/scheduler.h
-        )
+    scheduler.c
+    )
 
-target_include_directories(scheduler PUBLIC ${BITLOOM_ROOT}/bitloom-core/include)
-target_include_directories(scheduler PRIVATE ${BITLOOM_PROJECT_CONFIG})
+target_include_directories(scheduler PUBLIC ${BITLOOM_ROOT}/include)
+target_include_directories(scheduler PRIVATE ${BITLOOM_CONFIG})
+
